+++ src/runner/centralized_kb_hook.cpp
@@ -3,7 +3,6 @@
 #include <common/debug_control.h>
 #include <common/utils/winapi_error.h>
 #include <common/logger/logger.h>
-#include <common/interop/shared_constants.h>
 
 namespace CentralizedKeyboardHook
 {
@@ -23,30 +22,6 @@ namespace CentralizedKeyboardHook
     std::mutex mutex;
     HHOOK hHook{};
 
-    // To store information about handling pressed keys.
-    struct PressedKeyDescriptor
-    {
-        DWORD virtualKey; // Virtual Key code of the key we're keeping track of.
-        std::wstring moduleName;
-        std::function<bool()> action;
-        UINT_PTR idTimer; // Timer ID for calling SET_TIMER with.
-        UINT millisecondsToPress; // How much time the key must be pressed.
-        bool operator<(const PressedKeyDescriptor& other) const
-        {
-            // We'll use the virtual key as the real key, since looking for a hit with the key is done in the more time sensitive path (low level keyboard hook).
-            return virtualKey < other.virtualKey;
-        };
-    };
-    std::multiset<PressedKeyDescriptor> pressedKeyDescriptors;
-    std::mutex pressedKeyMutex;
-
-    // keep track of last pressed key, to detect repeated keys and if there are more keys pressed.
-    const DWORD VK_DISABLED = CommonSharedConstants::VK_DISABLED;
-    DWORD vkCodePressed = VK_DISABLED;
-
-    // Save the runner window handle for registering timers.
-    HWND runnerWindow;
-
     struct DestroyOnExit
     {
         ~DestroyOnExit()
@@ -54,30 +29,6 @@ namespace CentralizedKeyboardHook
             Stop();
         }
     } destroyOnExitObj;
-
-    // Handle the pressed key proc
-    void PressedKeyTimerProc(
-        HWND hwnd,
-        UINT /*message*/,
-        UINT_PTR idTimer,
-        DWORD /*dwTime*/)
-    {
-        std::multiset<PressedKeyDescriptor> copy;
-        {
-            // Make a copy, to look for the action to call.
-            std::unique_lock lock{ pressedKeyMutex };
-            copy = pressedKeyDescriptors;
-        }
-        for (const auto& it : copy)
-        {
-            if (it.idTimer == idTimer)
-            {
-                it.action();
-            }
-        }
-
-        KillTimer(hwnd, idTimer);
-    }
 
     LRESULT CALLBACK KeyboardHookProc(_In_ int nCode, _In_ WPARAM wParam, _In_ LPARAM lParam)
     {
@@ -92,50 +43,6 @@ namespace CentralizedKeyboardHook
         {
             // The new keystroke was generated from one of our actions. We should pass it along.
             return CallNextHookEx(hHook, nCode, wParam, lParam);
-        }
-
-        // Check if the keys are pressed.
-        if (!pressedKeyDescriptors.empty())
-        {
-            bool wasKeyPressed = vkCodePressed != VK_DISABLED;
-            // Hold the lock for the shortest possible duration
-            if ((wParam == WM_KEYDOWN || wParam == WM_SYSKEYDOWN))
-            {
-                if (!wasKeyPressed)
-                {
-                    // If no key was pressed before, let's start a timer to take into account this new key.
-                    std::unique_lock lock{ pressedKeyMutex };
-                    PressedKeyDescriptor dummy{ .virtualKey = keyPressInfo.vkCode };
-                    auto [it, last] = pressedKeyDescriptors.equal_range(dummy);
-                    for (; it != last; ++it)
-                    {
-                        SetTimer(runnerWindow, it->idTimer, it->millisecondsToPress, PressedKeyTimerProc);
-                    }
-                }
-                else if (vkCodePressed != keyPressInfo.vkCode)
-                {
-                    // If a different key was pressed, let's clear the timers we have started for the previous key.
-                    std::unique_lock lock{ pressedKeyMutex };
-                    PressedKeyDescriptor dummy{ .virtualKey = vkCodePressed };
-                    auto [it, last] = pressedKeyDescriptors.equal_range(dummy);
-                    for (; it != last; ++it)
-                    {
-                        KillTimer(runnerWindow, it->idTimer);
-                    }
-                }
-                vkCodePressed = keyPressInfo.vkCode;
-            }
-            if (wParam == WM_KEYUP || wParam == WM_SYSKEYUP)
-            {
-                std::unique_lock lock{ pressedKeyMutex };
-                PressedKeyDescriptor dummy{ .virtualKey = keyPressInfo.vkCode };
-                auto [it, last] = pressedKeyDescriptors.equal_range(dummy);
-                for (; it != last; ++it)
-                {
-                    KillTimer(runnerWindow, it->idTimer);
-                }
-                vkCodePressed = 0x100;
-            }
         }
 
         if ((wParam != WM_KEYDOWN) && (wParam != WM_SYSKEYDOWN))
@@ -194,17 +101,6 @@ namespace CentralizedKeyboardHook
         hotkeyDescriptors.insert({ .hotkey = hotkey, .moduleName = moduleName, .action = std::move(action) });
     }
 
-    void AddPressedKeyAction(const std::wstring& moduleName, const DWORD vk, const UINT milliseconds, std::function<bool()>&& action) noexcept
-    {
-        // Calculate a unique TimerID.
-        auto hash = std::hash<std::wstring>{}(moduleName); // Hash the module as the upper part of the timer ID.
-        const UINT upperId = hash & 0xFFFF;
-        const UINT lowerId = vk & 0xFFFF; // The key to press can be the lower ID.
-        const UINT timerId = upperId << 16 | lowerId;
-        std::unique_lock lock{ pressedKeyMutex };
-        pressedKeyDescriptors.insert({ .virtualKey = vk, .moduleName = moduleName, .action = std::move(action), .idTimer = timerId, .millisecondsToPress = milliseconds });
-    }
-
     void ClearModuleHotkeys(const std::wstring& moduleName) noexcept
     {
         Logger::trace(L"UnRegister hotkey action for {}", moduleName);
@@ -216,21 +112,6 @@ namespace CentralizedKeyboardHook
                 if (it->moduleName == moduleName)
                 {
                     it = hotkeyDescriptors.erase(it);
-                }
-                else
-                {
-                    ++it;
-                }
-            }
-        }
-        {
-            std::unique_lock lock{ pressedKeyMutex };
-            auto it = pressedKeyDescriptors.begin();
-            while (it != pressedKeyDescriptors.end())
-            {
-                if (it->moduleName == moduleName)
-                {
-                    it = pressedKeyDescriptors.erase(it);
                 }
                 else
                 {
@@ -269,8 +150,7 @@ namespace CentralizedKeyboardHook
         }
     }
 
-    void RegisterWindow(HWND hwnd) noexcept
+    void RegisterWindow(HWND /*hwnd*/) noexcept
     {
-        runnerWindow = hwnd;
     }
 }